pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "pygfs"
dynamic = ["version"]
description = "A Python package with async gRPC client"
authors = [{name = "Your Name", email = "your.email@example.com"}]
dependencies = [
    "grpcio>=1.35.0",
    "grpcio-tools>=1.35.0",
    "protobuf>=3.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-asyncio>=0.15.0",
    "black",
    "isort",
]

[tool.setuptools.packages.find]
where = ["src"]